--- public/docs/mentions_sources.txt
@@ -1,10 +1,0 @@
-
-MorisS are from Self-Translation of Mathematical Texts in Seventeenth-Century France:  The Cases of Pascal, Mersenne and Hérigone
-bib 1-25 are from: Axworthy, A. (2022). The Hybridization of Practical and Theoretical Geometry in the Sixteenth-Century Euclidean Tradition [OA]. Interdisciplinary Journal for the History of Ideas.
-bib 26-71 are from Raynaud, Dominique (ed.). 2015. Géométrie pratique: géomètres, ingénieurs et architectes XVIe-XVIIIe siècle. Besançon: Presses universitaires de Franche-Comté.
-bib 72-146 are from The Teaching of Elementary Geometry in the Seventeenth Century. Author(s): L. C. Karpinski and F. W. Kokomoor
-bib 147-157 are from Brioist, Pascal.. 2009. “ ‘Familiar Demonstrations in Geometry’: French and Italian Engineers and Euclid in the Sixteenth Century”. History of Science 47, no. 1: 1-26.
-bib 167-178 are from Oosterhoff, Richard. (2014). Idiotae, Mathematics, and Artisans: The Untutored Mind and the Discovery of Nature in the Fabrist Circle. Intellectual History Review. 24. 10.1080/17496977.2014.891180.
-bib 179-261 are "Manuals"
-
-Manuscripts with source "Lee" are from Lee, E. (2020). Visual Agency in Euclid's Elements: A Study of the Transmission of Visual Knowledge (Order No. 28103932). Available from ProQuest Dissertations & Theses Global; ProQuest One Academic; ProQuest One Literature. (2431809145).
