+++ public/docs/paratext_transcriptions.csv
@@ -1123,7 +1123,6 @@ Trinity College in Cambridge.
 Καθαρευούσης λογικῆς εἰσιν αἱ μαθηματικαὶ ἀλήθειαι. Hierocl.","LONDON,
 Printed by E. Redmayne in Jewen-street, 1705.",,
 "Kiel,_Hamburg_and_Leipzig_1697","Samuel Reyhers, JCti, Bey der Hoch-Fürstl. Holst. Universität Codic. und Mathematum Prof. In Teutscher Sprache vorgestellter EUCLIDES, Dessen VI. erste Bücher auf sonderbahre Art, Mit Algebraischen Zeichen, also eingerichtet sind, daß man derselben Beweise auch in andern Sprachen gebrauchen kan.","Kiel, Bey Benjamin Schillern, in Hamburg und Leipzig zu bekommen, Jm Jahr 1697.",,
-s.l_1650,,,,
 Oxford_1703,"ΕΥΚΛΕΙΔΟΥ
 ΤΑ ΣΩΖΟΜΕΝΑ.
 
@@ -1515,7 +1514,6 @@ PRINTED FOR BELL & BRADFUTE, AND
 G. G. & J. ROBINSON, LONDON.  
 M.DCC.XCV.",,
 Naples_1810,I primi sei libri degli elementi di Euclide ...,,,
-Paris_1814,,,,
 Nuremberg_1821,"Mathematisch-critische
 Bearbeitung  
 des  
@@ -7764,7 +7762,6 @@ Sed nunc melius studiorum consulate rebus.
 Si autem dialecticam artem cum modestia suscipitis: consequens est ut bonæ disciplinæ redeant omnes.  
 Bonas autem disciplinas morum probitas & vitæ decor comitatur omnis, omnisque virtus.  
 Quod summopere studiis optamus & impetramur.",,,
-bib-174,,,,
 bib-175,"Contenta.
 
 Primum volumen Contemplationum
