+++ public/docs/shelfmarks.csv
@@ -1,4 +1,4 @@
-key,volume,scan,title_page,frontispiece,annotations,shelf_mark,copyright
+key,volume,scan,title_page_img,frontispiece_img,annotations,shelf_mark,copyright
 ms_1,,http://digital.bodleian.ox.ac.uk/inquire/p/06cfa3b7-2aad-465e-88ac-0ebe3f2b5d13,,,,"Oxford, Bodleian, D'Orville, 301, 001-371",
 ms_2,,http://gallica.bnf.fr/ark:/12148/btv1b107229634/f5.image,,,,"Paris, Bibliothèque Nationale, Gr., 2466, 001-241v",
 ms_3,,http://digi.vatlib.it/mss/detail/178321,,,,"Vatican, Vaticana, Vat.gr., 0190, 003-292",
