+++ src/constants/index.ts
@@ -9,7 +9,6 @@ export const CSV_PATH_TRANSLATIONS = "/docs/translations.csv";
 export const CSV_PATH_SHELFMARKS = "/docs/shelfmarks.csv";
 export const CSV_PATH_CITIES = "/docs/cities.csv";
 export const CSV_PATH_DOTTED_LINES = "/docs/dotted_lines.csv";
-export const CSV_PATH_COPIES = "/docs/copies.csv";
 
 export const TILE_HEIGHT = 400;
 export const TILE_WIDTH = 400;
